refactor(user,weight): test move entities, dto
