+++ apps/api/src/modules/user/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Users } from './entities/users.entity';
+import { Users } from '@fithelper/api-interfaces';
 import { UserService } from './services/users.service';
 import { UsersController } from './user.controller';
