+++ apps/api/src/modules/weight/weight.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Users } from '../user/entities/users.entity';
-import { Weight } from './entities/weights.entity';
+import { Users } from '@fithelper/api-interfaces';
+import { Weight } from '@fithelper/api-interfaces';
 import { WeightService } from './services/weight.service';
 import { WeightController } from './weight.controller';
