+++ libs/api-interfaces/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/api-interfaces';
+export * from './lib/user';
+export * from './lib/weight';
