+++ libs/api-interfaces/src/lib/user/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './create-user.dto';
+export * from './output-user.dto';
+export * from './update.user.dto';
