+++ libs/api-interfaces/src/lib/user/index.ts
@@ -1,0 +1,2 @@
+export * from './dto';
+export * from './entities/users.entity';
