+++ libs/api-interfaces/src/lib/weight/dto/create-weight.dto.ts
