+++ libs/api-interfaces/src/lib/weight/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './create-weight.dto';
+export * from './output-weight.dto';
+export * from './update-weight.dto';
