+++ libs/api-interfaces/src/lib/weight/dto/output-weight.dto.ts
