+++ libs/api-interfaces/src/lib/weight/dto/update-weight.dto.ts
