+++ libs/api-interfaces/src/lib/weight/entities/weights.entity.ts
