+++ libs/api-interfaces/src/lib/weight/index.ts
@@ -1,0 +1,2 @@
+export * from './entities/weights.entity';
+export * from './dto';
